Fixes for palatalization and doubled consonants
So far, I've managed to change the doubling of "c" and "dz" into "{tinco}[bar-below][right-hook-below]" and "{ando}[bar-below][right-hook-below]" together with the misplaced [bar-below] in other doubled consonants, as well as fix the issue with the [double-dot-below] being placed under the preceding tengwa in the occurrence of palatalization (both for "nj" and "lj"), but I failed to come up with a solution to the problem of doubled of vowels such as "ii" in "priimek". That will be fixed after some testing of my own to find a solution.